fix(test): resolve PGlite harness via @revealui/db/testing graph
Move createTestDb into packages/db (export ./testing) so server, auth, ai,
and db tests import through the install graph instead of relative paths
into packages/test/src. Re-export from @revealui/test/utils for consumers
that already depend on the test package. Avoids turbo cycles that would
come from reverse-depending packages/test from auth/db.
