+++ apps/server/src/routes/__tests__/dunning-lifecycle.pglite.test.ts
@@ -24,14 +24,10 @@
  */
 
 import { accountEntitlements, accountMemberships, accounts } from '@revealui/db/schema';
+import { createTestDb, seedTestUser, type TestDb } from '@revealui/db/testing';
 import { eq } from 'drizzle-orm';
 import { Hono } from 'hono';
 import { afterAll, afterEach, beforeAll, describe, expect, it, vi } from 'vitest';
-import {
-  createTestDb,
-  seedTestUser,
-  type TestDb,
-} from '../../../../../packages/test/src/utils/drizzle-test-db.js';
 
 // ─── Mocks (before imports) ─────────────────────────────────────────────────
