+++ apps/server/src/routes/__tests__/entitlement-convergence.pglite.test.ts
@@ -19,13 +19,9 @@
  * every DB write, the idempotency state machine, and the saga run for real.
  */
 
+import { createTestDb, seedTestUser, type TestDb } from '@revealui/db/testing';
 import { eq } from 'drizzle-orm';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
-import {
-  createTestDb,
-  seedTestUser,
-  type TestDb,
-} from '../../../../../packages/test/src/utils/drizzle-test-db.js';
 
 // ─── Mocks (before imports) ─────────────────────────────────────────────────
