+++ packages/ai/src/__tests__/edge-cases.test.ts
@@ -17,8 +17,8 @@ import { createHash } from 'node:crypto';
 import type { AgentMemory } from '@revealui/contracts/agents';
 import type { Database } from '@revealui/db/client';
 import { nodeIdMappings } from '@revealui/db/schema';
+import { createTestDb, type TestDb } from '@revealui/db/testing';
 import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
-import { createTestDb, type TestDb } from '../../../test/src/utils/drizzle-test-db.js';
 import { NodeIdService } from '../memory/services/node-id-service.js';
 import { EpisodicMemory } from '../memory/stores/episodic-memory.js';
 import type { VectorMemoryService } from '../memory/vector/vector-memory-service.js';
