+++ packages/ai/src/__tests__/node-id-service.test.ts
@@ -1,9 +1,9 @@
 import { createHash } from 'node:crypto';
 import type { Database } from '@revealui/db/client';
 import { nodeIdMappings } from '@revealui/db/schema';
+import { createTestDb, type TestDb } from '@revealui/db/testing';
 import { eq } from 'drizzle-orm';
 import { afterAll, beforeAll, beforeEach, describe, expect, it } from 'vitest';
-import { createTestDb, type TestDb } from '../../../test/src/utils/drizzle-test-db.js';
 import { NodeIdService } from '../memory/services/node-id-service.js';
 
 /**
