+++ packages/test/package.json
@@ -50,6 +50,16 @@
       "types": "./dist/integration/test-database.d.ts",
       "import": "./dist/integration/test-database.js",
       "default": "./dist/integration/test-database.js"
+    },
+    "./utils": {
+      "types": "./dist/utils/index.d.ts",
+      "import": "./dist/utils/index.js",
+      "default": "./dist/utils/index.js"
+    },
+    "./utils/drizzle-test-db": {
+      "types": "./dist/utils/drizzle-test-db.d.ts",
+      "import": "./dist/utils/drizzle-test-db.js",
+      "default": "./dist/utils/drizzle-test-db.js"
     }
   },
   "main": "./dist/index.js",
